feat(tailscale): add local and deployment configurations for tailscale
Introduce new YAML configurations for Tailscale local and deployment setups. This includes the addition of `tailscale-local.yaml` and `tailscale-local-deployment.yaml` for local Tailscale configuration, as well as the migration of `tailscale-gluetun-sg.yaml` and `tailscale-gluetun-us.yaml` to deployment files `tailscale-gluetun-sg-deployment.yaml` and `tailscale-gluetun-us-deployment.yaml`. These changes aim to improve the scalability and maintainability of Tailscale configurations in the Kubernetes environment.